requirements.txt
streamlit
streamlit-chat
openai
requests
Pillow
langchain
feedparser
flake8
pytest
